+++ src/player/UserFactory.java
@@ -3,7 +3,7 @@ package player;
 public class UserFactory implements AbstractPlayerFactory {
 
     @Override
-    public User createPlayer(int intelligence){
+    public User createPlayer(double intelligence){
         User player = new User();
         return player;
     }
